Created new submodul for requests
Moved Ringorequest into ringo/lib/request/app.py and separated Featuretoggle
into its own file.
