+++ ringo/lib/request/app.py
@@ -1,6 +1,9 @@
 from pyramid.events import NewRequest
 from pyramid.decorator import reify
 
+from ringo.lib.request.featuretoggle import FeatureToggle
+
+
 class RingoRequest(object):
 
     """Ringo specific functions and properties directly attached to the
@@ -14,24 +17,10 @@ class RingoRequest(object):
         """Cached property to the FeatureToggle configuration."""
         return FeatureToggle(self.request.registry.settings)
 
-class FeatureToggle(object):
-
-    """FeatureToggle will give access to the configuration of feature
-    toggle configuration in the config file. Feature can be configured
-    usind *feature.name* config variables in the ini file. 
-
-    To enable a feature you the feature must be set to *true* Any other
-    value will evaluate to False which means the feature is not
-    enabled."""
-
-    def __init__(self, settings):
-        self.settings = settings
-
-    def __getattr__(self, name):
-        return self.settings.get("feature.%s" % name) == "true"
 
 def includeme(config):
     config.add_subscriber(add_ringo_request, NewRequest)
+
 
 def add_ringo_request(event):
     event.request.ringo = RingoRequest(event.request)
